requirements.txt
flask
requests
cffi
scrapyrt
gunicorn
bs4
tweepy
python-dotenv
psycopg2-binary
